fix: fix navigation
