+++ lib/features/candidates/candidates_list/presentation/candidate_tile.dart
@@ -1,5 +1,5 @@
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:employee_app_team/app/view/app.dart';
+import 'package:employee_app_team/routes/app_routes.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
